requirements.txt
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
certifi==2018.10.15
cffi==1.11.5
channels==2.1.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
daphne==2.2.2
Django==2.1.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
lxml==4.2.5
parsel==1.5.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-scrapyd-api==2.1.2
pytz==2018.6
queuelib==1.5.0
requests==2.20.0
Scrapy==1.5.1
scrapyd==1.2.0
service-identity==17.0.0
six==1.11.0
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24
w3lib==1.19.0
zope.interface==4.6.0
